Se mueve la funcion extract_emoji a util para reutilizarla
